prop-settings: rename react bridge for clarity
